+++ src/interrupt/timer.rs
@@ -1,6 +1,6 @@
 use x86_64::structures::idt::InterruptStackFrame;
 
-use crate::interrupts::{notify_end_of_interrupt, InterruptIndex};
+use crate::interrupt::{notify_end_of_interrupt, InterruptIndex};
 
 pub extern "x86-interrupt" fn handler(_frame: InterruptStackFrame) {
     // crate::print!(".");
